requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
asyncio==3.4.3
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
multidict==6.0.4
pyTelegramBotAPI==4.12.0
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.17
typing_extensions==4.7.0
urllib3==2.0.3
yarl==1.9.2
